microservices/auth-service/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
bcrypt==4.0.1
PyJWT==2.8.0
redis==4.6.0
